Move code out of table_metadata.hpp
